NowPlaying Message and Interactions

+++ src/services/music/dynamicMessage.ts
@@ -1,4 +1,5 @@
 import { Message, MessageCreateOptions, MessageEditOptions, MessagePayload, TextBasedChannel } from "discord.js";
+import ClassLogger from "../../utils/logger";
 
 export class DynamicMessage {
     /** Text channel in which the message will be sent. */
@@ -15,13 +16,15 @@ export class DynamicMessage {
 
     /* ==== METHODS ========================================================= */
     /** Updates message content - used for any new message or update. */
-    async setContent(content: any): Promise<DynamicMessage> {
+    setContent(content: any): DynamicMessage {
         this.content = content;
         return this;
     }
 
     /** Sends new message with current content. */
     async send() {
+        ClassLogger.trace("Entering DynamicMessage::send()");
+
         if(this.content) {
             this.message = await this.textChannel.send(this.content)
                 .catch();
@@ -29,22 +32,47 @@ export class DynamicMessage {
     }
 
     /** Deletes current message (if possible). */
-    async delete() {
+    async delete(): Promise<void> {
+        ClassLogger.trace("Entering DynamicMessage::delete()");
+
+        ClassLogger.trace(`Message: ${!!this.message} - Deletable: ${this.message?.deletable}`);
+
         if(this.message?.deletable) {
             await this.message.delete()
-                .catch();
+                .catch(e => ClassLogger.warn("delete(): ", e));
+
+            // Remove message from memory, not useful anymore.
+            // Also the "deletable" property is not updated; keeping the message
+            // would result in a message.delete() on an invalid message.
+            this.message = undefined;
         }
     }
 
     /** Deletes message (if possible) and sends new one with current content. */
     async resend() {
-        await this.delete();
-        await this.send();
+        ClassLogger.trace("Entering DynamicMessage::resend()");
+
+        // If the last message in the channel is the current message, update without deleting
+        let update: boolean = !!(
+            this.message &&
+            (await this.textChannel.messages.fetch({ limit: 1 })
+                .catch(_ => undefined))
+                ?.first()?.id === this.message.id
+        );
+        
+        if(update) {
+            await this.update();
+        } else {
+            await this.delete();
+            await this.send();
+        }
     }
 
     /** Updates current message (if possible) with current content.
      *  if not possible, the resend() method is called instead. */
     async update() {
+        ClassLogger.trace("Entering DynamicMessage::update()");
+
         if(this.content) {
             if(this.message?.editable) {
                 await this.message.edit(this.content)
@@ -58,10 +86,12 @@ export class DynamicMessage {
     /** Updates the text channel to be used and
      *  resends the message (if any) on the new one.  */
     async updateChannel(textChannel: TextBasedChannel) {
+        ClassLogger.trace("Entering DynamicMessage::updateChannel()");
+
         if(this.textChannel.id !== textChannel.id) {
             this.textChannel = textChannel;
             if(this.message?.deletable) {
-                this.resend();
+                await this.resend();
             }
         }
     }
